Write logs from s3backend test to disk
